+++ src/main/java/gildedrose/shop/ShopOutputBoundary.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface ShopOutputBoundary {
     void displayInventory(List<ItemResponse> inventory);
-    void displayBalance(int balance);
+    void displayBalance(float balance);
 }
